Revert using purposes check
